+++ test/tls/test_extensions.py
@@ -1,15 +1,17 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-import unittest
 try:
     from unittest import mock
 except ImportError:
     import mock
 
+from test.common.classes import TestLoggerBase
+
 from cryptoparser.tls.algorithm import TlsECPointFormat, TlsNextProtocolName, TlsProtocolName
 from cryptoparser.tls.ciphersuite import TlsCipherSuite
 from cryptoparser.tls.extension import (
+    TlsECPointFormatVector,
     TlsExtensionApplicationLayerProtocolNegotiation,
 )
 from cryptoparser.tls.subprotocol import (
@@ -23,7 +25,7 @@ from cryptolyzer.tls.client import L7ClientTlsBase
 from cryptolyzer.tls.extensions import AnalyzerExtensions
 
 
-class TestTlsExtensions(unittest.TestCase):
+class TestTlsExtensions(TestLoggerBase):
     @staticmethod
     def get_result(host, port, protocol_version=TlsProtocolVersionFinal(TlsVersion.TLS1_2), timeout=None, ip=None):
         analyzer = AnalyzerExtensions()
@@ -37,6 +39,9 @@ class TestTlsExtensions(unittest.TestCase):
 
         result = self.get_result('badssl.com', 443)
         self.assertEqual(set(result.next_protocols), set([TlsNextProtocolName.HTTP_1_1, ]))
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers application layer protocol "http/1.1"', log_lines)
+        self.assertIn('Server offers next protocol(s) "http/1.1"', log_lines)
 
     @mock.patch.object(
         L7ClientTlsBase, 'do_tls_handshake',
@@ -63,6 +68,9 @@ class TestTlsExtensions(unittest.TestCase):
 
         result = self.get_result('badssl.com', 443)
         self.assertEqual(set(result.application_layer_protocols), set([TlsProtocolName.HTTP_1_1]))
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers application layer protocol "http/1.1"', log_lines)
+        self.assertIn('Server offers next protocol(s) "http/1.1"', log_lines)
 
     @mock.patch.object(
         L7ClientTlsBase, 'do_tls_handshake',
@@ -118,68 +126,100 @@ class TestTlsExtensions(unittest.TestCase):
             set(result.compression_methods),
             set([TlsCompressionMethod.NULL, ])
         )
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers compression method(s) "NULL"', log_lines)
 
     def test_ec_point_formats(self):
         result = self.get_result('ecc256.badssl.com', 433)
-        self.assertFalse(
-            result.encrypt_then_mac_supported,
-            [
-                TlsECPointFormat.UNCOMPRESSED,
-                TlsECPointFormat.ANSIX962_COMPRESSED_PRIME,
-                TlsECPointFormat.ANSIX962_COMPRESSED_CHAR2,
-            ]
+        self.assertEqual(
+            result.ec_point_formats,
+            [TlsECPointFormat.UNCOMPRESSED, ]
         )
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers point format(s) "UNCOMPRESSED"', log_lines)
 
         result = self.get_result('www.cloudflare.com', 443)
-        self.assertFalse(
-            result.encrypt_then_mac_supported,
-            [TlsECPointFormat.UNCOMPRESSED, ]
-            )
+        self.assertEqual(
+            result.ec_point_formats,
+            TlsECPointFormatVector([TlsECPointFormat.UNCOMPRESSED, ])
+        )
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers point format(s) "UNCOMPRESSED"', log_lines)
 
     def test_encrypt_then_mac(self):
         result = self.get_result('tls-v1-0.badssl.com', 1010, TlsProtocolVersionFinal(TlsVersion.TLS1_0))
         self.assertFalse(result.encrypt_then_mac_supported)
+        log_lines = self.pop_log_lines()
+        self.assertNotIn('Server does not offer encrypt then MAC', log_lines)
+        self.assertNotIn('Server offers encrypt then MAC', log_lines)
 
         result = self.get_result('www.cloudflare.com', 443)
         self.assertFalse(result.encrypt_then_mac_supported)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server does not offer encrypt then MAC', log_lines)
 
         result = self.get_result('www.facebook.com', 443)
         self.assertTrue(result.encrypt_then_mac_supported)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers encrypt then MAC', log_lines)
 
         result = self.get_result('www.protonmail.com', 443)
         self.assertIsNone(result.encrypt_then_mac_supported)
+        log_lines = self.pop_log_lines()
+        self.assertNotIn('Server does not offer encrypt then MAC', log_lines)
+        self.assertNotIn('Server offers encrypt then MAC', log_lines)
 
     def test_extended_master_secret(self):
         result = self.get_result('tls-v1-2.badssl.com', 1012)
         self.assertFalse(result.extended_master_secret_supported)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server does not offer extended master secret', log_lines)
 
         result = self.get_result('www.cloudflare.com', 443)
         self.assertTrue(result.extended_master_secret_supported)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers extended master secret', log_lines)
 
     def test_clock_is_accurate(self):
         result = self.get_result('www.facebook.com', 443)
         self.assertFalse(result.clock_is_accurate)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server does not offer accurate clock', log_lines)
 
         result = self.get_result('www.cloudflare.com', 443)
         self.assertTrue(result.clock_is_accurate)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers accurate clock', log_lines)
 
     def test_renegotiation_info(self):
         result = self.get_result('www.deloton.com', 443)
         self.assertFalse(result.renegotiation_supported)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server does not offer renegotiation', log_lines)
 
         result = self.get_result('www.cloudflare.com', 443)
         self.assertTrue(result.renegotiation_supported)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers renegotiation', log_lines)
 
     def test_session_cache(self):
         result = self.get_result('www.github.com', 443)
         self.assertFalse(result.session_cache_supported)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server does not offer session cache', log_lines)
 
         result = self.get_result('www.cloudflare.com', 443)
         self.assertTrue(result.session_cache_supported)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers session cache', log_lines)
 
     def test_session_ticket(self):
         result = self.get_result('www.github.com', 443)
         self.assertFalse(result.session_ticket_supported)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server does not offer session ticket', log_lines)
 
         result = self.get_result('www.cloudflare.com', 443)
         self.assertTrue(result.session_ticket_supported)
+        log_lines = self.pop_log_lines()
+        self.assertIn('Server offers session ticket', log_lines)